LayoutTests/http/tests/xmlhttprequest/XMLHttpRequestException-expected.txt

Tests the properties of the exceptions related to XMLHttpRequest.

PASS e.toString() is "NetworkError (DOM Exception 19):  A network error occurred."
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) is not "[object DOMException]". Was "[object DOMExceptionPrototype]"
PASS e.constructor.toString() is "function DOMException() {
    [native code]
}"
PASS e.constructor is window.DOMException
PASS e.NETWORK_ERR is e.constructor.NETWORK_ERR
PASS e.NETWORK_ERR is 19
